requirements.txt
python-redmine
python-dotenv
Pillow
PyYAML